# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This documentation for Echo REST server.",
        "title": "Echo REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Dzung Tran",
            "url": "https://docs.api.com/support",
            "email": "support@api.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "XFirebaseBearer": []
                    }
                ],
                "description": "Get list user info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get list user info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of records should be returned",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wrapper.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domains.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    }
                }
            }
        },
        "/admin/users/{userId}": {
            "get": {
                "security": [
                    {
                        "XFirebaseBearer": []
                    }
                ],
                "description": "Get user info by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wrapper.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domains.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domains.User": {
            "description": "User account information",
            "type": "object",
            "properties": {
                "code": {
                    "description": "User last name",
                    "type": "string",
                    "example": "95a8d1aa-xxx-xxx-0c15d41"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "email@api.com"
                },
                "first_name": {
                    "description": "User first name",
                    "type": "string",
                    "example": "Dzung"
                },
                "id": {
                    "description": "User indetifier number",
                    "type": "integer",
                    "example": 1
                },
                "last_name": {
                    "description": "User last name",
                    "type": "string",
                    "example": "Tran"
                },
                "phone": {
                    "type": "string",
                    "example": "+84 0986415xxxx"
                },
                "status": {
                    "enum": [
                        "active",
                        "deactivated",
                        "banned"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domains.UserStatus"
                        }
                    ],
                    "example": "active"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domains.UserStatus": {
            "type": "string",
            "enum": [
                "active",
                "deactivated",
                "banned",
                "invited"
            ],
            "x-enum-varnames": [
                "UserStatusActive",
                "UserStatusDeactivated",
                "UserStatusBanned",
                "UserStatusInvited"
            ]
        },
...
        "wrapper.FailResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Example message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "wrapper.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "metadata": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "XFirebaseBearer": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. `Bearer jwt_token_string`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "XUserEmailAuth": {
            "description": "This method just enabled for local development",
            "type": "apiKey",
            "name": "X-User-Email",
            "in": "header"
        }
    }
}